backend/requirements.txt
    azure-functions
    requests
